Corrección tipo de dato moneda

+++ src/paquete/Usuario.java
@@ -6,14 +6,14 @@ public class Usuario {
 
 	private final String NOMBRE;
 	private final TipoAtraccion TEMATICA_FAVORITA;
-	private final int MONEDAS_INICIALES;
-	private int monedasDisponibles;
+	private final double MONEDAS_INICIALES;
+	private double monedasDisponibles;
 	private final double TIEMPO_INICIAL;
 	private double tiempoDisponible;
 	protected Itinerario itinerarioUsuario;
 	private ArrayList<Atraccion> listaAtracciones = new ArrayList<Atraccion>();
 
-	public Usuario(String nombre, TipoAtraccion tematica, int monedas, double tiempo) {
+	public Usuario(String nombre, TipoAtraccion tematica, double monedas, double tiempo) {
 		this.NOMBRE = nombre;
 		this.TEMATICA_FAVORITA = tematica;
 		this.MONEDAS_INICIALES = monedas;
@@ -31,7 +31,7 @@ public class Usuario {
 		return TEMATICA_FAVORITA;
 	}
 
-	public int getMonedasDisponibles() {
+	public double getMonedasDisponibles() {
 		return monedasDisponibles;
 	}
 
@@ -43,7 +43,7 @@ public class Usuario {
 		return listaAtracciones;
 	}
 	
-	public int monedasUtilizadas() {
+	public double monedasUtilizadas() {
 		return this.MONEDAS_INICIALES - this.monedasDisponibles;
 	}
